Issue #1202: Change gpi_set_signal_value_long(long) to gpi_set_signal_value_int(int32_t)
- vpi_put_value() is limited to 32-bit, so use int32_t instead of long.
- Fix 32-bit platform compilation warning: comparison of integer expressions of different signedness.

(cherry picked from commit 99079fa70b5f7b41bd68f52dd1ec9fc0e4e3b4b2)

+++ cocotb/share/lib/vhpi/VhpiImpl.h
@@ -195,7 +195,7 @@ public:
     long get_signal_value_long() override;
 
     using GpiSignalObjHdl::set_signal_value;
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
     int set_signal_value(double value, gpi_set_action_t action) override;
     int set_signal_value_str(std::string &value, gpi_set_action_t action) override;
     int set_signal_value_binstr(std::string &value, gpi_set_action_t action) override;
@@ -222,7 +222,7 @@ public:
 
 
     using GpiSignalObjHdl::set_signal_value;
-    int set_signal_value(long value, gpi_set_action_t action) override;
+    int set_signal_value(int32_t value, gpi_set_action_t action) override;
     int set_signal_value_binstr(std::string &value, gpi_set_action_t action) override;
 
     int initialise(std::string &name, std::string &fq_name) override;
